requirements.txt
beautifulsoup4==4.13.4
certifi==2025.4.26
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cssselect==1.3.0
feedfinder2==0.0.4
feedparser==6.0.11
filelock==3.18.0
idna==3.10
jieba3k==0.35.1
joblib==1.4.2
lxml==5.4.0
lxml_html_clean==0.4.2
newspaper3k==0.2.8
nltk==3.9.1
numpy==2.2.5
pandas==2.2.3
pillow==11.2.1
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-file==2.1.0
scikit-learn==1.6.1
scipy==1.15.2
sgmllib3k==1.0.0
six==1.17.0
soupsieve==2.7
textblob==0.19.0
threadpoolctl==3.6.0
tinysegmenter==0.3
tldextract==5.3.0
tqdm==4.67.1
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
